update README.md
Adding note to myself that I'm partway through updating the code on the DigitalOcean droplet to be able to use environment variables for the paths where the data storage will be in order to make this more portable across processing systems. Currently have the 'activate' file modified with correct environment variable contents, but haven't changed code.